+++ digiplan/static/js/elements.js
@@ -1,0 +1,2 @@
+export const resultsDropdown = document.getElementById("result_views");
+export const resultsTabs = document.getElementById("results-tabs");